Set Processes comm name to package name
When debugging audit logs, it is often helpful to
have the packages name. Bear in mind that the
package name is the right most chars up to
TASK_COMM_LEN. Although a possibility of truncation
exists, it can be useful for narrowing down the
possibilities.

This is on master as change-id:
Ia735b03f64352d155f8b96e1d47f48e6519b1d4f

Change-Id: I580f25dd3df5b2ade0b6fcf07ac6d91ccecad981