Add workarounds for apparent clang bugs
For reasons yet to be debugged, clang 3.4 can't resolve
Arena::kDefaultSize and kIndexUnresolved in the 2 locations.

The workaround is to copy in their values instead.
This should be reverted once clang is fixed (or an
issue in the code legitimately blocking clang from
seeing the symbols has been found and fixed).

Change-Id: Ic392de1829a44215299aee78dfe26378b0b3bef4
Signed-off-by: Bernhard Rosenkränzer <Bernhard.Rosenkranzer@linaro.org>