Implement GenInlinedSqrt for x86
Implemented this in the same manner as is done on ARM.
If the hardware instruction gives an error, call sqrt().

Why is it necessary to invoke sqrt?  Is setting errno the reason?

I am not happy with the code that I wrote.  I don't understand why
it is necessary to call ClobberSReg to get the code to work.  Can
someone help me write this in a better fashion?

Tested manually using host mode.

Change-Id: I082aa20041c933ae5fc78f12ddf491d1c775c683
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>