[WIP] Enable background compaction.
When the process state changes to a state which does not
perceives jank, we copy from the main free-list backed
allocation space to the bump pointer space and enable the
semispace allocator.

When we transition back to foreground, we copy back to a
free-list backed space.

Bugs to fix:
Race conditions where threads are suspended inside of the entrypoints code.

TODO improvements: Recycle mem-maps which are unused in the current state.

WORK IN PROGRESS

Bug: 8981901

Change-Id: Ib235d03f45548ffc08a06b8ae57bf5bada49d6f3