Fix JDWP IsCollected command handling of null.
We used to rely on ObjectRegistry::Get returning NULL to detect null object id
(in which case we return an INVALID_OBJECT error). The issue is this method
may also return NULL for a valid object id if the corresponding object has
been collected. In this case, we must not return any JDWP error and the result
of the command should be 'true' meaning the object identified by the given id
has been collected.

This CL avoids this confusion by moving out the handling of 'null' object id.

Change-Id: I013f3759bac7cf5b1e9ee3e8c0d1bb2a11f1b654