[WIP] Remove blacklist
... Not ready for review ...

Experiment removal of class initialization blacklist using transactions. The
idea is to log every change made to objects and internal runtime state so we
can revert them when we can't initialize an image class.

Here are the main changes:
- Transaction implementation (undo version) keeping track of every original
field value, original array values (per index) and every intern string table
modification.
- Transaction attached to the runtime so it can be accessed from everywhere.
- Support transactional and non-transactional in mirror::Object, mirror::Array,
mirror::ObjectArray and InternTable routines. Transactional versions log
changes into the current transaction. Non transactional versions ensure no
transaction is pending.
- Templatize mirror::Object and mirror::Array interfaces to control transaction
mode.

Known issues/workarounds:
- We check whether transaction mode is enabled or not in each field/array write
routine to not miss anything. This may have negative impact on performance. I
did not evaluate this for the moment but this is the next thing on my list once
everything works correctly. We should mimit the use of transactional routines
to reduce overhead there.
- Need to preserve Object::class and Array::length fields when aborting
transaction to not mess up with GC.
- Force monothreaded classes initialization: we need to ensure a compilation
thread entering in transaction mode has exclusive access to the runtime. This
can be implemented in a separate CL. Note we only do this when creating image
so applications are not impacted.
- Force GC run after we processed every class initialization so we can prune
objects created in aborted transaction so we never visit them when creating
the image. Note we only do this when creating image so applications are not
impacted.

Bug: 9676614
Change-Id: I221910a9183a5ba6c2b99a277f5a5a68bc69b5f9