Restore thread state after running checkpoints in signal catcher.
Previously, we were restoring the thread state before running any
pending checkpoints. This caused an issue with profiler checkpoints
since these use a ScopedObjectAccess. Since the SIGQUIT handler has
exclusive access to the mutator lock, this meant that we deadlocked
when we tried to gain shared access to the mutator lock in the
ScopedThreadStateChange if the old_state was any state other than
kRunnable.

Bug: 12490061
Change-Id: I5f922693a4bf5d18cf9d22f1f28a7a04b6bfc077