Don't DecrementDisableGC if not JNI_COMMIT, delete pin table.
The pin table was unused and would not work with the current moving
collector. It is also not necessary to use the pin table to hold
roots since these objects should be held live by the reference
tables. If someone were to delete the jobject while still holding a
pointer to array data, there would be no way of releasing the
primitive array.

Change-Id: I8f6beef18c03cff93387a6967f32a8891348642c