Use macros to remove CFI information for MAC OS/X
Ian asked me to change the CFI information in the x86 .S files to allow
them to be compiled out when building on MAC OS/X, since they aren't
supported in the assembler. This compiles on Linux, and is untested
under MAC OS/X because I don't have access to a MAC system.

Change-Id: Id0d83578e9b93b0f6aea5e9b8ddf2815f0dcd8aa
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>