[WIP] Support Direct Method/Type access for X86
Thumb generates code to optimize calls to methods within core.oat.
Implement this for X86 as well, using the new literal pool pc-relative
access handling.

Fix some incorrect return locations for long inlines.

This will be replaced with a version that directly generates mov
instructions and relative calls shortly, avoiding use of the literal
pool.  That will remove the need for analysis of the INVOKE
instructions. This version served to let me figure out what was going
on.

Change-Id: I1907bdfc7574f3d0aa76c7fad13dc537acdf1ed3
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>