[WIP] Implicit null checks in compiler and runtime
This work in progress adds the ability to use SEGV signals
to throw NullPointerException exceptions from Java code rather
than having the compiler generate explicit comparisons and
branches.

It does this by using sigaction to trap SIGSEGV and when triggered
makes sure it's in compiled code and if so, sets the return
address to the entry point to throw the exception.  If the user
program traps SIGSEGV and doesn't pass it on, this won't
work.  If we don't handle the SIGSEGV we log a fatal error
as there can be nothing else to trap it.

The default is OFF.  To switch it on either:
1. Pass -implicit-null-checks to dalvikvm
2. setprop dalvik.vm.implicit_null_checks 1

It only works for ARM right now.

Also adds 2 new option args to oatdump:

--dump:raw_mapping_table
--dump:raw_gc_map

They should be self explanitory.

Bug: 13121132
Change-Id: If743849138162f3c7c44a523247e413785677370