Implicit null/suspend checks
This adds the ability to use SEGV signals
to throw NullPointerException exceptions from Java code rather
than having the compiler generate explicit comparisons and
branches.  It does this by using sigaction to trap SIGSEGV and when triggered
makes sure it's in compiled code and if so, sets the return
address to the entry point to throw the exception.

It also uses this signal mechanism to determine whether to check
for thread suspension.  Instead of the compiler generating calls
to a function to check for threads being suspended, the compiler
will now load indirect via an address in the TLS area.  To trigger
a suspend, the contents of this address are changed from something
valid to 0.  A SIGSEGV will occur and the handler will check
for a valid instruction pattern before invoking the thread
suspension check code.

This version of the code revert the requirement to modify
bionic to prevent user's calling sigaction and overriding
our use of it.  Any user code that traps SIGSEGV will break
this.

The default is OFF.  To switch it on either:
1. Pass -implicit-checks to dalvikvm
2. setprop dalvik.vm.implicit_checks 1

It only works for ARM right now.

Bug: 13121132
Change-Id: If743849138162f3c7c44a523247e413785677370